rework into expression module
